Add test for locking4rep
